Migrate to Spring Boot 2.7 and Java 11
Migrated eml-cts project to Spring Boot 2.7 and Java 11, making only minor changes to the POM.
This is the first step before updating to Spring Boot 3.0 and Java 21. The jump to Spring Boot 3.0
is likely to require more substantive changes to the POM. This commit serves as a base before that
final jump.