Migrate to Spring Boot 3.0.13 and Java 21
Migrated eml-cts project to Spring Boot 3 and Java 21. Replacing all 'javax' (as in
'javax.persistence' and 'javax.validation') to 'jakarta' was the only major change made to the codebase.
So more minor updates may still be warranted to move up to Spring Boot 3.2 (the most recent release
of Spring Boot) and to handle code that has been deprecated.