Migrate to Spring Boot 3.1
Migrated to Spring Boot 3.1. The Hibernate Dialect was changed from MySQL5 to MySQL due to the
version specific MySQL5 dialects being removed. Hibernate 6 (the version of Hibernate being used by this
Spring Boot) version deprecated MySQL5 dialect and other version-specific dialects a while ago, and it
was removed on this version of Spring Boot. According to the Hibernate migration guide, Hibernate
should be able to detect the version of the database and act accordingly (the reason for the deprecation).
Please refer to the migration guide for more details: https://docs.jboss.org/hibernate/orm/6.0/migration-guide/migration-guide.html#version-specific-and-spatial-dialects