Migrate to Spring Boot 3.2 and Java 21
Migrated to Spring Boot 3.2 and Java 21. Most of the changes are located in the POM, but some changes
were made to the code base (such as changing `javax` to `jakarta` for all imports throughout the code base).

There should be no changes to the functionality of the code as all changes were in response to changes in
dependencies and deprecations. This commit is effectively a recreation of a different attempt to migrate
to Java 21 but this time ensuring that the gitignore fully ignored IDE-specific files.