Create new classes for TenderBase
Created new classes that will be used for the new TenderBase class of the March 2024 Draft Specification.
Besides the inclusion of these classes and the renaming of old ones, the functionality of the codebase
remains unaltered. The inclusion of these new classes is in preparation for adjusting the codebase
so that it fits the March 2024 spec.