Use new TenderBase class from new spec
Modified the codebase so that it works with the new TenderBase class as specified by the March 2024
Specification.

TenderDetail may change in the near future, as Stream Tenders have not be incorporated into the codebase, and
that may require non-trivial changes to the codebase. Changes to the Tender facet, such as the payloads
are still needed.