Add initial changes to EiCreateTenderPayload to comply with Spec
Added all the new required fields as specified by the March 2024 specification.
Made a few assumptions about the new fields, such as there only be one market segment of the market,
thus the segmentID is one by default. The assumptions were chosen to make the initial changes as minimal
as possible, and to make it easy for future developers to change the functionality if needed or if the
assumptions were wrong.